LPS-200228 Rename and move to components folder
